Reduce title to just Cluster URL for every individual cluster under every Panel in a Grafana Dashboard
On Grafana Dashboards for:
1> Single Arch Build
2> Multiarch Build &
3> RPM builds

For every Panel, the title of the panel repeats for every Cluster. Since there are multiple clusters per panel and hence multiple windows per panel, the name/URL of the actual CLUSTER gets hidden. We already know the title from the Panel, so we can avoid the same repeatation of title name again before the actual CLUSTER name/URL.
This PR aims to improve that.